% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connection_Function.R
\name{connectToNeo4j}
\alias{connectToNeo4j}
\title{Connect to Neo4j graph database}
\usage{
connectToNeo4j(url, user, db)
}
\arguments{
\item{url}{the Neo4j server url.}

\item{user}{The username of the Neo4j server ("neo4j" is the default user).}

\item{db}{The database name.}
}
\value{
A connection object to the Neo4j Server.
}
\description{
Connect to Neo4j graph database
}
\examples{
\dontrun{
connection = connectToNeo4j("http://localhost:7474", "neo4j", "pathways")
connection$ping()
connection = connectToNeo4j()
}

}
