% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Creation_Functions.R
\name{createNode}
\alias{createNode}
\title{Create a node with its property keys in Neo4j database}
\usage{
createNode(connection, node_label, property_keys = NULL)
}
\arguments{
\item{connection}{The Neo4j connection object.}

\item{node_label}{The node label.}

\item{property_keys}{The node property keys, in the form of dataframe.}
}
\value{
The result from the Neo4j call.
}
\description{
Create a single node with its property keys, passed as arguments, in a Neo4j database.
}
\examples{
\dontrun{
#Get the dataframe of nodes to be created
nodes = prepareNodesDataframeToCreate(wntBiopax,wntmully)

#Create all nodes in the Neo4j database
for(row in 1:nrow(nodes)){
 node_property_keys = nodes[row, names(nodes) != "label"]

 na_cells = which(is.na(node_property_keys))
 names_not_na = names(node_property_keys)[-na_cells]
 node_property_keys = as.data.frame(node_property_keys[,-na_cells])
 colnames(node_property_keys) = names_not_na

 createNode(connection,nodes[row,'label'],node_property_keys)
}
}
}
