% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Creation_Functions.R
\name{createRelationship}
\alias{createRelationship}
\title{Create a relationship with its property keys in Neo4j database}
\usage{
createRelationship(
  connection,
  started_node_label,
  started_node_property,
  started_node_property_value,
  end_node_label,
  end_node_property,
  end_node_property_value,
  relationship_label,
  relationship_property_keys = NULL
)
}
\arguments{
\item{connection}{The Neo4j connection object.}

\item{started_node_label}{The started node label.}

\item{started_node_property}{The started node property key name to be identified by.}

\item{started_node_property_value}{The started node property key value to be found within.}

\item{end_node_label}{The end node label.}

\item{end_node_property}{The end node property key name to be identified by.}

\item{end_node_property_value}{The end node property key value to be found within.}

\item{relationship_label}{The relationship label.}

\item{relationship_property_keys}{The relationship property keys, in the form of dataframe.}
}
\value{
The result from the Neo4j call.
}
\description{
Create a single relationship with its property keys, passed as arguments, in a Neo4j database.
}
\examples{
\dontrun{
#Get the dataframe of relationships to be created
relationships = prepareRelationshipsDataframeToCreate(wntmully,nodes)

#Create all nodes in the Neo4j database
for(row in 1:nrow(relationships)){
 relationship_property_keys = relationships[row, !names(relationships) \%in\%
 c("label", "V1 label", "V2 label", "V1", "V2")]

 na_cells = which(is.na(relationship_property_keys))
 names_not_na = names(relationship_property_keys)[-na_cells]
 relationship_property_keys = as.data.frame(relationship_property_keys[,-na_cells])
 colnames(relationship_property_keys) = names_not_na

 createRelationship(connection, relationships[row,'V1_label'], 'name',
                    relationships[row,'V1'], relationships[row,'V2_label'],
                      'name', relationships[row,'V2'], relationships[row,'label'],
                         relationship_property_keys)
}
}
}
