% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Getter_Functions.R
\name{getLabel}
\alias{getLabel}
\title{Get the nodes on a layer in a Neo4j graph}
\usage{
getLabel(connection, label, property_key)
}
\arguments{
\item{connection}{The Neo4j connection object.}

\item{label}{The name of the label.}

\item{property_key}{The node property key name to be extracted.}
}
\value{
A list of nodes having the label, with their property key values.
}
\description{
Get the nodes on a layer in a Neo4j graph
}
\examples{
\dontrun{
#To get the list of Proteins' names
result = getLabel(connection, "Protein", "name")
}
}
