% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Getter_Functions.R
\name{getLabelsCount}
\alias{getLabelsCount}
\title{Get the number of labels in a Neo4j graph}
\usage{
getLabelsCount(connection)
}
\arguments{
\item{connection}{The Neo4j connection object.}
}
\value{
The labels count in case of success, otherwise the result from the Neo4j call.
}
\description{
Get the number of labels in a Neo4j graph
}
\examples{
\dontrun{
#To get the labels count
result = getLabelsCount(connection)
}
}
