% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Preparation_Functions.R
\name{prepareNodesDataframeToCreate}
\alias{prepareNodesDataframeToCreate}
\title{Prepare the nodes dataframe}
\usage{
prepareNodesDataframeToCreate(biopax, g)
}
\arguments{
\item{biopax}{The biopax object.}

\item{g}{The mully graph.}
}
\value{
A dataframe of nodes ready to be created in Neo4j database.
}
\description{
Prepare a dataframe containing the nodes of a mully graph, extracted from getNodesAttribues() function, to be created in Neo4j database. The dataframe contains the nodes' label and their property keys.
}
\examples{
\dontrun{
nodes = prepareNodesDataframeToCreate(wntBiopax,wntmully)
}
}
