% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Preparation_Functions.R
\name{prepareRelationshipsDataframeToCreate}
\alias{prepareRelationshipsDataframeToCreate}
\title{Prepare the relationships dataframe}
\usage{
prepareRelationshipsDataframeToCreate(g, nodes)
}
\arguments{
\item{g}{The mully graph.}

\item{nodes}{The nodes dataframe, containing the nodes' labels}
}
\value{
A dataframe of relationships ready to be created in Neo4j database.
}
\description{
Prepare a dataframe containing the relationships of a mully graph, extracted from getEdgesAttribues() function, to be created in Neo4j database. The dataframe contains the relationships' label, extremities nodes' labels, relationships' labels and their property keys.
}
\examples{
\dontrun{
#Get the dataframe of nodes to be created
relationships = prepareRelationshipsDataframeToCreate(wntmully,nodes)
}
}
