% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Removal_Functions.R
\name{removeAllNodesOfLabel}
\alias{removeAllNodesOfLabel}
\title{Delete all nodes of the same label}
\usage{
removeAllNodesOfLabel(connection, label)
}
\arguments{
\item{connection}{The Neo4j connection object.}

\item{label}{The nodes label to be deleted.}
}
\value{
The result from the Neo4j call.
}
\description{
Delete all the nodes of the same label.
}
\examples{
\dontrun{
#To delete all labels
for(label in wntmully$layers$Name){
 removeAllNodesOfLabel(connection,label)
}

#To delete drugs nodes
removeAllNodesOfLabel(connection,"drugs")
}
}
