% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Removal_Functions.R
\name{removeNode}
\alias{removeNode}
\title{Delete a node in the Neo4j graph}
\usage{
removeNode(connection, property_key, property_key_value)
}
\arguments{
\item{connection}{The Neo4j connection object.}

\item{property_key}{The node property key name to be identified by.}

\item{property_key_value}{The node property key value to be found within.}
}
\value{
The result from the Neo4j call.
}
\description{
Delete a node or a set of nodes from a graph
}
\examples{
\dontrun{
#To remove the node of name Protein1
removeNode(connection, "name", "Protein1")
}
}
