% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Preparation_Functions.R
\name{splitDBIDs}
\alias{splitDBIDs}
\title{Split the databases' ids' of the nodes extracted from the function getNodesAttributes()}
\usage{
splitDBIDs(biopax, nodes_df)
}
\arguments{
\item{biopax}{The biopax object.}

\item{nodes_df}{The nodes dataframe.}
}
\value{
A dataframe containing the nodes with their ids in each database respectively.
}
\description{
Split the databases' ids' of the nodes extracted from the function getNodesAttributes()
}
\examples{
\dontrun{
nodes_dataframe = splitDBIDs(biopax, nodes_dataframe)
}
}
