% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Store_Mully_Function.R
\name{storeMully}
\alias{storeMully}
\title{Save a mully graph in the Neo4j database}
\usage{
storeMully(connection, biopax, mully)
}
\arguments{
\item{connection}{The Neo4j connection object.}

\item{biopax}{The biopax object.}

\item{mully}{The mully graph.}
}
\value{
A boolean value: TRUE if the mully graph has been saved in the Neo4j server. Otherwise FALSE.
}
\description{
Save a mully graph in the Neo4j database
}
\examples{
\dontrun{
storeMully(connection, wntBiopax, wntmully)
}
}
